@@ -7,14 +7,12 @@ import { AlertifyService } from '../_services/alertify.service';
   providedIn: 'root'
 })
 export class AuthGuard implements CanActivate {
-  constructor(private authService: AuthService, private router: Router,
-    private alertify: AlertifyService) {}
+  constructor(private authService: AuthService, private router: Router, private alertify: AlertifyService) {}
 
   canActivate(): boolean {
     if (this.authService.loggedIn()) {
       return true;
     }
-
     this.alertify.error('You shall not pass!!!');
     this.router.navigate(['/home']);
     return false;
