@@ -1,6 +1,6 @@
 import {Injectable} from '@angular/core';
-import { CanDeactivate } from '@angular/router';
 import { MemberEditComponent } from '../members/member-edit/member-edit.component';
+import { CanDeactivate } from '@angular/router';
 
 @Injectable()
 export class PreventUnsavedChanges implements CanDeactivate<MemberEditComponent> {
