@@ -1,13 +1,13 @@
 export interface Message {
    id: number;
-   senderId: number;
-   senderKnownAs: string;
-   senderPhotoUrl: string;
-   recipientId: number;
-   recipientKnownAs: string;
-   recipientPhotoUrl: string;
-   content: string;
-   isRead: boolean;
-   dateRead: Date;
-   messageSent: Date;
+  senderId: number;
+  senderKnownAs: string;
+  senderPhotoUrl: string;
+  recipientId: number;
+  recipientKnownAs: string;
+  recipientPhotoUrl: string;
+  content: string;
+  isRead: boolean;
+  dateRead: Date;
+  messageSent: Date;
 }
