@@ -3,10 +3,9 @@ export interface Pagination {
     itemsPerPage: number;
     totalItems: number;
     totalPages: number;
-}
+  }
 
-
-export class PaginatedResult<T> {
+  export class PaginatedResult<T> {
     result: T;
     pagination: Pagination;
-}
+  }
