@@ -1,24 +1,28 @@
-import {Injectable} from '@angular/core';
-import {User} from '../_models/user';
-import {Resolve, Router, ActivatedRouteSnapshot} from '@angular/router';
-import { UserService } from '../_services/user.service';
+import { Injectable } from '@angular/core';
+import { User } from '../_models/user';
+import { Resolve, Router, ActivatedRouteSnapshot } from '@angular/router';
 import { AlertifyService } from '../_services/alertify.service';
+import { UserService } from '../_services/user.service';
 import { Observable, of } from 'rxjs';
 import { catchError } from 'rxjs/operators';
 import { AuthService } from '../_services/auth.service';
 
 @Injectable()
 export class MemberEditResolver implements Resolve<User> {
-    constructor(private userService: UserService, private router: Router,
-        private alertify: AlertifyService, private authService: AuthService) {}
+  constructor(
+    private userService: UserService,
+    private router: Router,
+    private alertify: AlertifyService,
+    private authService: AuthService
+  ) {}
 
-    resolve(route: ActivatedRouteSnapshot): Observable<User> {
-        return this.userService.getUser(this.authService.decodedToken.nameid).pipe(
-            catchError(error => {
-                this.alertify.error('Problem retrieving your data');
-                this.router.navigate(['/members']);
-                return of(null);
-            })
-        );
-    }
+  resolve(route: ActivatedRouteSnapshot): Observable<User> {
+    return this.userService.getUser(this.authService.decodedToken.nameid).pipe(
+      catchError(error => {
+        this.alertify.error('Problem retrieving your data');
+        this.router.navigate(['/members']);
+        return of(null);
+      })
+    );
+  }
 }
