@@ -1,26 +1,29 @@
-import {Injectable} from '@angular/core';
-import {User} from '../_models/user';
-import {Resolve, Router, ActivatedRouteSnapshot} from '@angular/router';
-import { UserService } from '../_services/user.service';
+import { Injectable } from '@angular/core';
+import { User } from '../_models/user';
+import { Resolve, Router, ActivatedRouteSnapshot } from '@angular/router';
 import { AlertifyService } from '../_services/alertify.service';
+import { UserService } from '../_services/user.service';
 import { Observable, of } from 'rxjs';
 import { catchError } from 'rxjs/operators';
 
 @Injectable()
 export class MemberListResolver implements Resolve<User[]> {
-    pageNumber = 1;
-    pageSize = 5;
+  pageNumber = 1;
+  pageSize = 5;
 
-    constructor(private userService: UserService, private router: Router,
-        private alertify: AlertifyService) {}
+  constructor(
+    private userService: UserService,
+    private router: Router,
+    private alertify: AlertifyService
+  ) {}
 
-    resolve(route: ActivatedRouteSnapshot): Observable<User[]> {
-        return this.userService.getUsers(this.pageNumber, this.pageSize).pipe(
-            catchError(error => {
-                this.alertify.error('Problem retrieving data');
-                this.router.navigate(['/home']);
-                return of(null);
-            })
-        );
-    }
+  resolve(route: ActivatedRouteSnapshot): Observable<User[]> {
+    return this.userService.getUsers(this.pageNumber, this.pageSize).pipe(
+      catchError(error => {
+        this.alertify.error('Problem retrieving data');
+        this.router.navigate(['/']);
+        return of(null);
+      })
+    );
+  }
 }
