@@ -1,29 +1,39 @@
-import {Injectable} from '@angular/core';
-import {Resolve, Router, ActivatedRouteSnapshot} from '@angular/router';
+import { Injectable } from '@angular/core';
+import { Resolve, Router, ActivatedRouteSnapshot } from '@angular/router';
 import { UserService } from '../_services/user.service';
 import { AlertifyService } from '../_services/alertify.service';
 import { Observable, of } from 'rxjs';
 import { catchError } from 'rxjs/operators';
-import { AuthService } from '../_services/auth.service';
 import { Message } from '../_models/message';
+import { AuthService } from '../_services/auth.service';
 
 @Injectable()
 export class MessagesResolver implements Resolve<Message[]> {
-    pageNumber = 1;
-    pageSize = 5;
-    messageContainer = 'Unread';
+  pageNumber = 1;
+  pageSize = 5;
+  messageContainer = 'Unread';
 
-    constructor(private userService: UserService, private router: Router,
-                private authService: AuthService, private alertify: AlertifyService) {}
+  constructor(
+    private userService: UserService,
+    private router: Router,
+    private alertify: AlertifyService,
+    private authService: AuthService
+  ) {}
 
-    resolve(route: ActivatedRouteSnapshot): Observable<Message[]> {
-        return this.userService.getMessages(this.authService.decodedToken.nameid,
-                this.pageNumber, this.pageSize, this.messageContainer).pipe(
-            catchError(error => {
-                this.alertify.error('Problem retrieving messages');
-                this.router.navigate(['/home']);
-                return of(null);
-            })
-        );
-    }
+  resolve(route: ActivatedRouteSnapshot): Observable<Message[]> {
+    return this.userService
+      .getMessages(
+        this.authService.decodedToken.nameid,
+        this.pageNumber,
+        this.pageSize,
+        this.messageContainer
+      )
+      .pipe(
+        catchError(error => {
+          this.alertify.error('Problem retrieving messages');
+          this.router.navigate(['/home']);
+          return of(null);
+        })
+      );
+  }
 }
