@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-declare let alertify: any;
+import * as alertify from 'alertifyjs';
 
 @Injectable({
   providedIn: 'root'
@@ -8,10 +8,11 @@ export class AlertifyService {
   constructor() {}
 
   confirm(message: string, okCallback: () => any) {
-    alertify.confirm(message, function(e) {
+    alertify.confirm(message, (e: any) => {
       if (e) {
         okCallback();
-      } else {}
+      } else {
+      }
     });
   }
 
