@@ -3,7 +3,7 @@ import { User } from '../_models/user';
 import { Pagination, PaginatedResult } from '../_models/pagination';
 import { AuthService } from '../_services/auth.service';
 import { UserService } from '../_services/user.service';
-import { ActivatedRoute } from '@angular/router';
+import { ActivatedRoute } from '../../../node_modules/@angular/router';
 import { AlertifyService } from '../_services/alertify.service';
 
 @Component({
@@ -16,9 +16,12 @@ export class ListsComponent implements OnInit {
   pagination: Pagination;
   likesParam: string;
 
-  constructor(private authService: AuthService,
-              private userService: UserService,
-              private route: ActivatedRoute, private alertify: AlertifyService) { }
+  constructor(
+    private authService: AuthService,
+    private userService: UserService,
+    private route: ActivatedRoute,
+    private alertify: AlertifyService
+  ) {}
 
   ngOnInit() {
     this.route.data.subscribe(data => {
@@ -29,13 +32,22 @@ export class ListsComponent implements OnInit {
   }
 
   loadUsers() {
-    this.userService.getUsers(this.pagination.currentPage, this.pagination.itemsPerPage, null, this.likesParam)
-    .subscribe((res: PaginatedResult<User[]>) => {
-      this.users = res.result;
-      this.pagination = res.pagination;
-    }, error => {
-      this.alertify.error(error);
-    });
+    this.userService
+      .getUsers(
+        this.pagination.currentPage,
+        this.pagination.itemsPerPage,
+        null,
+        this.likesParam
+      )
+      .subscribe(
+        (res: PaginatedResult<User[]>) => {
+          this.users = res.result;
+          this.pagination = res.pagination;
+        },
+        error => {
+          this.alertify.error(error);
+        }
+      );
   }
 
   pageChanged(event: any): void {
