@@ -1,8 +1,8 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { Message } from 'src/app/_models/message';
-import { UserService } from 'src/app/_services/user.service';
-import { AuthService } from 'src/app/_services/auth.service';
-import { AlertifyService } from 'src/app/_services/alertify.service';
+import { Message } from '../../_models/message';
+import { UserService } from '../../_services/user.service';
+import { AuthService } from '../../_services/auth.service';
+import { AlertifyService } from '../../_services/alertify.service';
 import { tap } from 'rxjs/operators';
 
 @Component({
@@ -15,8 +15,11 @@ export class MemberMessagesComponent implements OnInit {
   messages: Message[];
   newMessage: any = {};
 
-  constructor(private userService: UserService, private authService: AuthService,
-              private alertify: AlertifyService) { }
+  constructor(
+    private userService: UserService,
+    private authService: AuthService,
+    private alertify: AlertifyService
+  ) {}
 
   ngOnInit() {
     this.loadMessages();
@@ -24,30 +27,42 @@ export class MemberMessagesComponent implements OnInit {
 
   loadMessages() {
     const currentUserId = +this.authService.decodedToken.nameid;
-    this.userService.getMessageThread(this.authService.decodedToken.nameid, this.recipientId)
+    this.userService
+      .getMessageThread(this.authService.decodedToken.nameid, this.recipientId)
       .pipe(
         tap(messages => {
           for (let i = 0; i < messages.length; i++) {
-            if (messages[i].isRead === false && messages[i].recipientId === currentUserId) {
+            if (
+              messages[i].isRead === false &&
+              messages[i].recipientId === currentUserId
+            ) {
               this.userService.markAsRead(currentUserId, messages[i].id);
             }
           }
         })
       )
-      .subscribe(messages => {
-        this.messages = messages;
-      }, error => {
-        this.alertify.error(error);
-      });
+      .subscribe(
+        messages => {
+          this.messages = messages;
+        },
+        error => {
+          this.alertify.error(error);
+        }
+      );
   }
+
   sendMessage() {
     this.newMessage.recipientId = this.recipientId;
-    this.userService.sendMessage(this.authService.decodedToken.nameid, this.newMessage)
-      .subscribe((message: Message) => {
-        this.messages.unshift(message);
-        this.newMessage.content = '';
-      }, error => {
-        this.alertify.error(error);
-      });
+    this.userService
+      .sendMessage(this.authService.decodedToken.nameid, this.newMessage)
+      .subscribe(
+        (message: Message) => {
+          this.messages.unshift(message);
+          this.newMessage.content = '';
+        },
+        error => {
+          this.alertify.error(error);
+        }
+      );
   }
 }
