@@ -14,10 +14,14 @@ import { ActivatedRoute } from '@angular/router';
 export class MessagesComponent implements OnInit {
   messages: Message[];
   pagination: Pagination;
-  messageContainer: 'Unread';
+  messageContainer = 'Unread';
 
-  constructor(private userService: UserService, private authService: AuthService,
-              private alertify: AlertifyService, private route: ActivatedRoute) { }
+  constructor(
+    private userService: UserService,
+    private authService: AuthService,
+    private route: ActivatedRoute,
+    private alertify: AlertifyService
+  ) {}
 
   ngOnInit() {
     this.route.data.subscribe(data => {
@@ -27,26 +31,44 @@ export class MessagesComponent implements OnInit {
   }
 
   loadMessages() {
-    this.userService.getMessages(this.authService.decodedToken.nameid,
-      this.pagination.currentPage, this.pagination.itemsPerPage, this.messageContainer)
-      .subscribe((res: PaginatedResult<Message[]>) => {
-        this.messages = res.result;
-        this.pagination = res.pagination;
-      }, error => {
-        this.alertify.error(error);
-      });
+    this.userService
+      .getMessages(
+        this.authService.decodedToken.nameid,
+        this.pagination.currentPage,
+        this.pagination.itemsPerPage,
+        this.messageContainer
+      )
+      .subscribe(
+        (res: PaginatedResult<Message[]>) => {
+          this.messages = res.result;
+          this.pagination = res.pagination;
+        },
+        error => {
+          this.alertify.error(error);
+        }
+      );
   }
 
   deleteMessage(id: number) {
-    this.alertify.confirm('Are you sure you would like to delete this message!', () => {
-      this.userService.deleteMessage(id, this.authService.decodedToken.nameid)
-      .subscribe(() => {
-        this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
-        this.alertify.success('Message has been deleted');
-      }, error => {
-        this.alertify.error('Failed to delete message');
-      });
-    });
+    this.alertify.confirm(
+      'Are you sure you want to delete this message?',
+      () => {
+        this.userService
+          .deleteMessage(id, this.authService.decodedToken.nameid)
+          .subscribe(
+            () => {
+              this.messages.splice(
+                this.messages.findIndex(m => m.id === id),
+                1
+              );
+              this.alertify.success('Message has been deleted');
+            },
+            error => {
+              this.alertify.error('Failed to delete the message');
+            }
+          );
+      }
+    );
   }
 
   pageChanged(event: any): void {
