@@ -12,26 +12,32 @@ export class NavComponent implements OnInit {
   model: any = {};
   photoUrl: string;
 
-  constructor(public authService: AuthService, private alertify: AlertifyService,
-      private router: Router) { }
+  constructor(
+    public authService: AuthService,
+    private alertify: AlertifyService,
+    private router: Router
+  ) {}
 
   ngOnInit() {
     this.authService.currentPhotoUrl.subscribe(photoUrl => this.photoUrl = photoUrl);
   }
 
   login() {
-    this.authService.login(this.model).subscribe(next => {
-      this.alertify.success('Logged in successfully');
-    }, error => {
-      this.alertify.error(error);
-    }, () => {
-      this.router.navigate(['/members']);
-    });
+    this.authService.login(this.model).subscribe(
+      next => {
+        this.alertify.success('Logged in successfully');
+      },
+      error => {
+        this.alertify.error(error);
+      },
+      () => {
+        this.router.navigate(['/members']);
+      }
+    );
   }
 
   loggedIn() {
-    const token = localStorage.getItem('token');
-    return !!token;
+    return this.authService.loggedIn();
   }
 
   logout() {
@@ -42,5 +48,4 @@ export class NavComponent implements OnInit {
     this.alertify.message('logged out');
     this.router.navigate(['/home']);
   }
-
 }
