@@ -1,4 +1,4 @@
-import {Routes} from '@angular/router';
+import { Routes } from '@angular/router';
 import { HomeComponent } from './home/home.component';
 import { MemberListComponent } from './members/member-list/member-list.component';
 import { MessagesComponent } from './messages/messages.component';
@@ -14,21 +14,39 @@ import { ListsResolver } from './_resolvers/lists.resolver';
 import { MessagesResolver } from './_resolvers/messages.resolver';
 
 export const appRoutes: Routes = [
-    {path: 'home', component: HomeComponent},
-    {
-        path: '',
-        runGuardsAndResolvers: 'always',
-        canActivate: [AuthGuard],
-        children: [
-            {path: 'members', component: MemberListComponent,
-                resolve: {users: MemberListResolver}},
-            {path: 'members/:id', component: MemberDetailComponent,
-                resolve: {user: MemberDetailResolver}},
-            {path: 'member/edit', component: MemberEditComponent,
-                resolve: {user: MemberEditResolver}, canDeactivate: [PreventUnsavedChanges]},
-            {path: 'messages', component: MessagesComponent, resolve: {messages: MessagesResolver}},
-            {path: 'lists', component: ListsComponent, resolve: {users: ListsResolver}},
-        ]
-    },
-    {path: '**', redirectTo: 'home', pathMatch: 'full'},
+  { path: '', component: HomeComponent },
+  {
+    path: '',
+    runGuardsAndResolvers: 'always',
+    canActivate: [AuthGuard],
+    children: [
+      {
+        path: 'members',
+        component: MemberListComponent,
+        resolve: { users: MemberListResolver }
+      },
+      {
+        path: 'members/:id',
+        component: MemberDetailComponent,
+        resolve: { user: MemberDetailResolver }
+      },
+      {
+        path: 'member/edit',
+        component: MemberEditComponent,
+        resolve: { user: MemberEditResolver },
+        canDeactivate: [PreventUnsavedChanges]
+      },
+      {
+        path: 'messages',
+        component: MessagesComponent,
+        resolve: { messages: MessagesResolver }
+      },
+      {
+        path: 'lists',
+        component: ListsComponent,
+        resolve: { users: ListsResolver }
+      }
+    ]
+  },
+  { path: '**', redirectTo: '', pathMatch: 'full' }
 ];
