@@ -1,147 +1,146 @@
-using System.Collections.Generic;
 using System;
-using System.Security.Claims;
-using System.Threading.Tasks;
-using AutoMapper;
-using DatingApp.API.Data;
-using DatingApp.API.Dtos;
-using DatingApp.API.Helpers;
-using DatingApp.API.Models;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
+ using System.Collections.Generic;
+ using System.Security.Claims;
+ using System.Threading.Tasks;
+ using AutoMapper;
+ using DatingApp.API.Data;
+ using DatingApp.API.Dtos;
+ using DatingApp.API.Helpers;
+ using DatingApp.API.Models;
+ using Microsoft.AspNetCore.Authorization;
+ using Microsoft.AspNetCore.Mvc;
 
-namespace DatingApp.API.Controllers
-{
-    [ServiceFilter(typeof(LogUserActivity))]
-    [Authorize]
-    [Route("api/users/{userId}/[controller]")]
-    [ApiController]
-    public class MessagesController : ControllerBase
-    {
-        private readonly IDatingRepository _repo;
-        private readonly IMapper _mapper;
+  namespace DatingApp.API.Controllers
+ {
+     [ServiceFilter(typeof(LogUserActivity))]
+     [Authorize]
+     [Route("api/users/{userId}/[controller]")]
+     [ApiController]
+     public class MessagesController : ControllerBase
+     {
+         private readonly IDatingRepository _repo;
+         private readonly IMapper _mapper;
+         public MessagesController(IDatingRepository repo, IMapper mapper)
+         {
+             _mapper = mapper;
+             _repo = repo;
+         }
 
-        public MessagesController(IDatingRepository repo, IMapper mapper)
-        {
-            _repo = repo;
-            _mapper = mapper;
-        }
+          [HttpGet("{id}", Name = "GetMessage")]
+         public async Task<IActionResult> GetMessage(int userId, int id)
+         {
+             if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
+                 return Unauthorized();
 
-        [HttpGet("{id}", Name="GetMessage")]
-        public async Task<IActionResult> GetMessage(int userId, int id)
-        {
-            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-                return Unauthorized();
+              var messageFromRepo = await _repo.GetMessage(id);
 
-            var messageFromRepo = await _repo.GetMessage(id);
+              if (messageFromRepo == null)
+                 return NotFound();
 
-            if (messageFromRepo == null) NotFound();
+              return Ok(messageFromRepo);
+         }
 
-            return Ok(messageFromRepo);
-            
-        }
+          [HttpGet]
+         public async Task<IActionResult> GetMessagesForUser(int userId, 
+             [FromQuery]MessageParams messageParams)
+         {
+             if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
+                 return Unauthorized();
 
-        [HttpGet]
-        public async Task<IActionResult> GetMessageForUser(int userId, [FromQuery]MessageParams messageParams)
-        {
-            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-                return Unauthorized();
+              messageParams.UserId = userId;
 
-            messageParams.UserId = userId;
+              var messagesFromRepo = await _repo.GetMessagesForUser(messageParams);
 
-            var messagesFromRepo = await _repo.GetMessagesForUser(messageParams);
+              var messages = _mapper.Map<IEnumerable<MessageToReturnDto>>(messagesFromRepo);
 
-            var messages = _mapper.Map<IEnumerable<MessageToReturnDto>>(messagesFromRepo);
+              Response.AddPagination(messagesFromRepo.CurrentPage, messagesFromRepo.PageSize, 
+                 messagesFromRepo.TotalCount, messagesFromRepo.TotalPages);
 
-            Response.AddPagination(messagesFromRepo.CurrentPage, messagesFromRepo.PageSize, 
-                    messagesFromRepo.TotalCount, messagesFromRepo.TotalPages);
+              return Ok(messages);
+         }
 
-            return Ok(messages);
-        }
+          [HttpGet("thread/{recipientId}")]
+         public async Task<IActionResult> GetMessageThread(int userId, int recipientId)
+         {
+             if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
+                 return Unauthorized();
 
-        [HttpGet("thread/{recipientId}")]
-        public async Task<IActionResult> GetMessageThread(int userId, int recipientId)
-        {
-            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-                return Unauthorized();
-            
-            var messageFromRepo = await _repo.GetMessagesThread(userId, recipientId);
+              var messagesFromRepo = await _repo.GetMessageThread(userId, recipientId);
 
-            var messageThread = _mapper.Map<IEnumerable<MessageToReturnDto>>(messageFromRepo);
+              var messageThread = _mapper.Map<IEnumerable<MessageToReturnDto>>(messagesFromRepo);
 
-            return Ok(messageThread);
-        }
+              return Ok(messageThread);
+         }
 
+          [HttpPost]
+         public async Task<IActionResult> CreateMessage(int userId, MessageForCreationDto messageForCreationDto)
+         {
+             var sender = await _repo.GetUser(userId);
 
-        [HttpPost]
-        public async Task<IActionResult> CreateMessage(int userId, MessageForCreationDto messageForCreationDto)
-        {
-            var sender = await _repo.GetUser(userId);
-            if (sender.Id != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-                return Unauthorized();
+              if (sender.Id != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
+                 return Unauthorized();
 
-            messageForCreationDto.SenderId = userId;
+              messageForCreationDto.SenderId = userId;
 
-            var recipient = await _repo.GetUser(messageForCreationDto.RecipientId);
+              var recipient = await _repo.GetUser(messageForCreationDto.RecipientId);
 
-            if (recipient == null)
-                return BadRequest("Could not find user");
-            
-            var message = _mapper.Map<Message>(messageForCreationDto);
+              if (recipient == null)
+                 return BadRequest("Could not find user");
 
-            _repo.Add(message);
+              var message = _mapper.Map<Message>(messageForCreationDto);
 
-            if (await _repo.SaveAll()) {
-                var messageToReturn = _mapper.Map<MessageToReturnDto>(message);
-                return CreatedAtRoute("Get Message", new {id = message.Id}, messageToReturn);
-            }
-                
+              _repo.Add(message);
 
-            throw new Exception("Creating the meassage failed save");
-        }
+              if (await _repo.SaveAll())
+             {
+                 var messageToReturn = _mapper.Map<MessageToReturnDto>(message);
+                 return CreatedAtRoute("GetMessage", 
+                     new {userId, id = message.Id}, messageToReturn);
+             }
 
-        [HttpPost("{id}")]
-        public async Task<IActionResult> DeleteMessage(int id, int userId)
-        {
-            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-                return Unauthorized();
+              throw new Exception("Creating the message failed on save");
+         }
 
-            var messageFromRepo = await _repo.GetMessage(id);
+          [HttpPost("{id}")]
+         public async Task<IActionResult> DeleteMessage(int id, int userId)
+         {
+             if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
+                 return Unauthorized();  
 
-            if (messageFromRepo.SenderId == userId)
-                messageFromRepo.SenderDeleted = true;
+              var messageFromRepo = await _repo.GetMessage(id);
 
-            if (messageFromRepo.RecipientId == userId)
-                messageFromRepo.RecipientDeleted = true;
+              if (messageFromRepo.SenderId == userId)
+                 messageFromRepo.SenderDeleted = true;
 
-            if (messageFromRepo.SenderDeleted && messageFromRepo.RecipientDeleted)
-                _repo.Delete(messageFromRepo);
+              if (messageFromRepo.RecipientId == userId)
+                 messageFromRepo.RecipientDeleted = true;
 
-            if (await _repo.SaveAll())
-                return NoContent();
-            
-            throw new Exception("Error deleting the message");
+              if (messageFromRepo.SenderDeleted && messageFromRepo.RecipientDeleted)
+                 _repo.Delete(messageFromRepo);
 
-        }
+              if (await _repo.SaveAll())
+                 return NoContent();
 
-         [HttpPost("{id}/read")]
+              throw new Exception("Error deleting the message");
+         }
+
+          [HttpPost("{id}/read")]
          public async Task<IActionResult> MarkMessageAsRead(int userId, int id)
          {
              if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-                return Unauthorized();
-            
-            var message = await _repo.GetMessage(id);
+                 return Unauthorized();  
 
-            if (message.RecipientId != userId)
-                return Unauthorized();
+              var message = await _repo.GetMessage(id);
 
-            message.IsRead = true;
-            message.DateRead = DateTime.Now;
+              if (message.RecipientId != userId)
+                 return Unauthorized();
 
-            await _repo.SaveAll();
+              message.IsRead = true;
+             message.DateRead = DateTime.Now;
 
-            return NoContent();
-           
+              await _repo.SaveAll();
+
+              return NoContent();
          }
-    }
-}
+     }
+ } 
