@@ -33,19 +33,19 @@ namespace DatingApp.API.Data
                 {
                     if (computedHash[i] != passwordHash[i]) return false;
                 }
-                return true;
             }
+            return true;
         }
 
         public async Task<User> Register(User user, string password)
         {
-            CreatePasswordHash(password, out byte[] passwordHash, out byte[] passwordSalt);
+            byte[] passwordHash, passwordSalt;
+            CreatePasswordHash(password, out passwordHash, out passwordSalt);
 
             user.PasswordHash = passwordHash;
             user.PasswordSalt = passwordSalt;
 
             await _context.Users.AddAsync(user);
-
             await _context.SaveChangesAsync();
 
             return user;
@@ -57,7 +57,7 @@ namespace DatingApp.API.Data
             {
                 passwordSalt = hmac.Key;
                 passwordHash = hmac.ComputeHash(System.Text.Encoding.UTF8.GetBytes(password));
-            } 
+            }
         }
 
         public async Task<bool> UserExists(string username)
