@@ -33,12 +33,12 @@ namespace DatingApp.API.Data
 
             modelBuilder.Entity<Message>()
             .HasOne(u => u.Sender)
-            .WithMany(m => m.MessagesSent)
+            .WithMany(m => m.MessagesSent )
             .OnDelete(DeleteBehavior.Restrict);
 
             modelBuilder.Entity<Message>()
             .HasOne(u => u.Recipient)
-            .WithMany(m => m.MessagesRecieved)
+            .WithMany(m => m.MessagesReceived)
             .OnDelete(DeleteBehavior.Restrict);
         }
     }
