@@ -9,22 +9,21 @@ namespace DatingApp.API.Data
     {
         public static void SeedUsers(DataContext context)
         {
-            if (!context.Users.Any())
+             if (!context.Users.Any())
             {
-                var userData =
-                    System.IO.File.ReadAllText("Data/UserSeedData.json");
+                var userData = System.IO.File.ReadAllText("Data/UserSeedData.json");
                 var users = JsonConvert.DeserializeObject<List<User>>(userData);
                 foreach (var user in users)
                 {
-                    CreatePasswordHash("password",
-                    out byte[] passwordHash,
-                    out byte[] passwordSalt);
+                    byte[] passwordHash, passwordSalt;
+                    CreatePasswordHash("password", out passwordHash, out passwordSalt);
 
                     user.PasswordHash = passwordHash;
                     user.PasswordSalt = passwordSalt;
                     user.Username = user.Username.ToLower();
-                    context.Users.Add (user);
+                    context.Users.Add(user);
                 }
+
                 context.SaveChanges();
             }
         }
@@ -38,13 +37,7 @@ namespace DatingApp.API.Data
             using (var hmac = new System.Security.Cryptography.HMACSHA512())
             {
                 passwordSalt = hmac.Key;
-                passwordHash =
-                    hmac
-                        .ComputeHash(System
-                            .Text
-                            .Encoding
-                            .UTF8
-                            .GetBytes(password));
+                passwordHash = hmac.ComputeHash(System.Text.Encoding.UTF8.GetBytes(password));
             }
         }
     }
