@@ -14,7 +14,7 @@ namespace DatingApp.API.Helpers
             var resultContext = await next();
 
             var userId = int.Parse(resultContext.HttpContext.User
-                .FindFirst(ClaimTypes.NameIdentifier).Value);
+               .FindFirst(ClaimTypes.NameIdentifier).Value);
             var repo = resultContext.HttpContext.RequestServices.GetService<IDatingRepository>();
             var user = await repo.GetUser(userId);
             user.LastActive = DateTime.Now;
