@@ -11,7 +11,7 @@ namespace DatingApp.API.Helpers
         public int CurrentPage { get; set; }
         public int TotalPages { get; set; }
         public int PageSize { get; set; }
-        public int TotalCount { get; set; }   
+        public int TotalCount { get; set; }
 
         public PagedList(List<T> items, int count, int pageNumber, int pageSize)
         {
@@ -20,13 +20,13 @@ namespace DatingApp.API.Helpers
             CurrentPage = pageNumber;
             TotalPages = (int)Math.Ceiling(count / (double)pageSize);
             this.AddRange(items);
-        } 
+        }
 
-        public static async Task<PagedList<T>> CreateAsync(IQueryable<T> source, int pageNumber, int pageSize)
+        public static async Task<PagedList<T>> CreateAsync(IQueryable<T> source,
+           int pageNumber, int pageSize)
         {
             var count = await source.CountAsync();
-            var items = await source.Skip((pageNumber -1) * pageSize).Take(pageSize).ToListAsync();
-
+            var items = await source.Skip((pageNumber - 1) * pageSize).Take(pageSize).ToListAsync();
             return new PagedList<T>(items, count, pageNumber, pageSize);
         }
     }
