@@ -4,7 +4,7 @@ namespace DatingApp.API.Models
     {
         public int LikerId { get; set; }
         public int LikeeId { get; set; }
-        public User Liker { get; set; }
-        public User Likee { get; set; }
+        public virtual User Liker { get; set; }
+        public virtual User Likee { get; set; }
     }
 }
